+++ valhalla/baldr/graphconstants.h
@@ -338,6 +338,7 @@ inline std::string to_string(Use u) {
       {static_cast<uint8_t>(Use::kPath), "path"},
       {static_cast<uint8_t>(Use::kPedestrian), "pedestrian"},
       {static_cast<uint8_t>(Use::kBridleway), "bridleway"},
+      {static_cast<uint8_t>(Use::kPedestrianCrossing), "pedestrian_crossing"},
       {static_cast<uint8_t>(Use::kRestArea), "rest_area"},
       {static_cast<uint8_t>(Use::kServiceArea), "service_area"},
       {static_cast<uint8_t>(Use::kOther), "other"},
@@ -360,10 +361,12 @@ inline std::string to_string(Use u) {
 enum class TaggedValue : uint8_t { // must start at 1 due to nulls
   kLayer = 1,
   kPronunciation = 2,
+  kBssInfo = 3,
   // we used to have bug when we encoded 1 and 2 as their ASCII codes, but not actual 1 and 2 values
   // see https://github.com/valhalla/valhalla/issues/3262
   kTunnel = static_cast<uint8_t>('1'),
   kBridge = static_cast<uint8_t>('2'),
+
 };
 
 enum class PronunciationAlphabet : uint8_t {